basic file reading and writing (no input validation)
