+++ sr/joke1.txt
