+++ sr/joke2.txt
