+++ sr/joke3.txt
